Fix type detection for models
Do not use model.url for determining model type. Was previously
used as a fallback if model.type does not exist however in
Backbone >= 0.5.3 model.url is a function and cannot be used
for this purpose anymore.

If a type attribute is not set in the model data, and there is
no model.type, then it simply will not have a type attribute.

The watchList now falls back to the collection url if the model
type is not set.